added phong lighting

+++ project/Project/Project/Light.h
@@ -8,8 +8,10 @@ class Light
 public:
 	Light();
 	Light(GLfloat red, GLfloat green, GLfloat blue, GLfloat a,
-		GLfloat xDir, GLfloat yDir, GLfloat zDir, GLfloat dIntensity);
-	void use(GLfloat aLoc, GLfloat cLoc, GLfloat dILoc, GLfloat dLoc);
+	      GLfloat xDir, GLfloat yDir, GLfloat zDir, GLfloat dIntensity);
+
+	void use(GLuint aLoc, GLuint cLoc, GLuint dILoc, GLuint dLoc);
+	
 	~Light();
 
 private:
